add trademark for canvass print outs

+++ apps/system/src/employee/employee.service.ts
@@ -123,7 +123,7 @@ export class EmployeeService {
 			}
 		})
 
-		if(!!input.signature_src) {
+		if(!!input.signature_src && !!existingItem.signature_src && existingItem.signature_src.trim() !== '') {
 			this.deleteFiles([existingItem.signature_src])
 			console.log('previous file deleted');
 		}
